Move integration add-on helpers to init
